pyproject.toml
[tool.poetry]
name = "django-react-notes"
version = "0.1.0"
description = ""
authors = ["rich <rkba2@outlook.com>"]
readme = "README.md"
packages = [{include = "django_react_notes"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
pymongo = "^4.7.3"
dnspython = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
